feature creep again
added a joke difficulty overhaul that basically makes the game easier on higher difficulties.
It was created because people were complaining about long assault waves, short assault breaks, inhuman reaction time of enemies, etc.

It was a mess to balance, I'm not gonna it seriously hence why it's labeled as a joke.

removed groupaistatestreet due to game crash

fixed crew bonus and weapon rebalance settings

todo: remove groupaistatebesiege if the implementation was ineffective.